+++ src/test/unit/relin-keys.test.js
@@ -39,7 +39,9 @@ describe('RelinKeys', () => {
     expect(item).toHaveProperty('getIndex')
     expect(item).toHaveProperty('hasKey')
     expect(item).toHaveProperty('save')
+    expect(item).toHaveProperty('saveArray')
     expect(item).toHaveProperty('load')
+    expect(item).toHaveProperty('loadArray')
     expect(item).toHaveProperty('copy')
     expect(item).toHaveProperty('clone')
     expect(item).toHaveProperty('move')
@@ -117,6 +119,13 @@ describe('RelinKeys', () => {
     expect(spyOn).toHaveBeenCalledWith()
     expect(typeof str).toBe('string')
   })
+  test('It should save to an array', () => {
+    const item = RelinKeysObject()
+    const spyOn = jest.spyOn(item, 'saveArray')
+    const array = item.saveArray()
+    expect(spyOn).toHaveBeenCalledWith()
+    expect(array.constructor).toBe(Uint8Array)
+  })
   test('It should load from a string', () => {
     const item = keyGenerator.genRelinKeys()
     const newItem = RelinKeysObject()
@@ -125,6 +134,15 @@ describe('RelinKeys', () => {
     newItem.load(context, str)
     expect(spyOn).toHaveBeenCalledWith(context, str)
     expect(newItem.save()).toEqual(str)
+  })
+  test('It should load from a typed array', () => {
+    const item = keyGenerator.genRelinKeys()
+    const newItem = RelinKeysObject()
+    const array = item.saveArray()
+    const spyOn = jest.spyOn(newItem, 'loadArray')
+    newItem.loadArray(context, array)
+    expect(spyOn).toHaveBeenCalledWith(context, array)
+    expect(newItem.saveArray()).toEqual(array)
   })
   test('It should fail to load from a string', () => {
     const newItem = RelinKeysObject()
@@ -138,6 +156,76 @@ describe('RelinKeys', () => {
     expect(spyOn).toHaveBeenCalledWith(
       context,
       'XqEAASUAAAAAAAAAAAAAAHicY2CgCHywj1vIwCCBRQYAOAcCRw=='
+    )
+  })
+  test('It should fail to load from a Uint8Array', () => {
+    const newItem = RelinKeysObject()
+    const spyOn = jest.spyOn(newItem, 'loadArray')
+    expect(() =>
+      newItem.loadArray(
+        context,
+        Uint8Array.from([
+          93,
+          161,
+          0,
+          1,
+          27,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          120,
+          156,
+          99,
+          103,
+          128,
+          0,
+          0,
+          0,
+          64,
+          0,
+          8
+        ])
+      )
+    ).toThrow()
+    expect(spyOn).toHaveBeenCalledWith(
+      context,
+      Uint8Array.from([
+        93,
+        161,
+        0,
+        1,
+        27,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        120,
+        156,
+        99,
+        103,
+        128,
+        0,
+        0,
+        0,
+        64,
+        0,
+        8
+      ])
     )
   })
   test('It should copy another instance', () => {