+++ CHANGES.md
@@ -3,6 +3,12 @@
 See [Microsoft's Change log](https://github.com/microsoft/SEAL/blob/master/Changes.md)
 for more details on each SEAL version change.
 
+## Version 4.1.5
+
+Feat:
+- Added `saveArray` and `loadArray` functions to all serializable components. These methods allow the objects to save
+ and load from a Uint8Array containing binary data.
+ 
 ## Version 4.1.4
 
 Feat: